qPCR Experiment for the Amplification of adk Using the Bio-Rad iQ5 thermo cycler

A quantitative real-time PCR of adk was performed.

data("C316.amp")

Format

A data frame with 40 observations on the following 97 variables. The first column ("Cycle") contains the number of cycles and consecutive columns contain the replicates ("A01" to "H12").

Details

adk was amplified in the Bio-Rad iQ5. The change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

gDNA: 28.43 ng/microL DNA concentration, 260/280 ratio= 1.96

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: 10 minutes at 95 degrees Celsius 40 x 30 seconds at 95 degrees Celsius 45 seconds at 59 degrees Celsius 45 seconds at 68 degrees Celsius

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

data(C316.amp)
str(C316.amp)

#> 'data.frame':	40 obs. of  98 variables:
#>  $ X          : logi  NA NA NA NA NA NA ...
#>  $ Well.Cycles: num  0.558 1.558 2.558 3.558 4.558 ...
#>  $ A01        : num  92.7 70.2 47.4 12.7 -16.1 ...
#>  $ A02        : num  6.82 -20.15 69.12 17.95 -10.45 ...
#>  $ A03        : num  92.14 62.83 25.56 4.23 -16.27 ...
#>  $ A04        : num  48.05 41.07 21.12 3.68 -8.75 ...
#>  $ A05        : num  45.31 30.62 15.48 5.52 -9.13 ...
#>  $ A06        : num  18.34 16.56 10.55 10.37 -8.28 ...
#>  $ A07        : num  19.627 18.858 -0.133 -1.535 -4.253 ...
#>  $ A08        : num  20.23 14.2 15.22 6.22 -2.49 ...
#>  $ A09        : num  31.33 26.12 19.64 8.43 -2.6 ...
#>  $ A10        : num  16.95 9.32 29.44 8.42 14.89 ...
#>  $ A11        : num  60.47 39.376 17.431 0.493 -14.715 ...
#>  $ A12        : num  74.59 53.36 31.48 16.43 -1.59 ...
#>  $ B01        : num  36.8 14.23 13.95 16.45 3.95 ...
#>  $ B02        : num  74.89 -1.36 -17.56 28.95 23.71 ...
#>  $ B03        : num  18.01 6.11 3.26 30.18 9.36 ...
#>  $ B04        : num  29.98 18.9 17.7 4.34 -6.67 ...
#>  $ B05        : num  42.17 30.471 10.847 0.717 -7.385 ...
#>  $ B06        : num  13.09 -24.39 27.07 17.17 6.95 ...
#>  $ B07        : num  21.03 22.59 13.9 6.68 -5.86 ...
#>  $ B08        : num  31.52 21.61 9.13 3.73 -7 ...
#>  $ B09        : num  25.52 24.58 14.38 4.12 -4.73 ...
#>  $ B10        : num  74.14 63.33 -3.25 -11.67 -15.84 ...
#>  $ B11        : num  -10.9 -31.3 -41.3 -31.3 25.4 ...
#>  $ B12        : num  65.12 46.21 23.43 6.76 -7.64 ...
#>  $ C01        : num  73.6 53 34.5 14.3 -10.1 ...
#>  $ C02        : num  62.62 25.68 64.97 6.68 -21.16 ...
#>  $ C03        : num  56.8 150.3 19.5 -32.8 -34.2 ...
#>  $ C04        : num  88.83 68.93 39.25 6.2 -8.58 ...
#>  $ C05        : num  38.93 23.38 13.35 5.49 -7.5 ...
#>  $ C06        : num  30.827 23.221 8.486 -0.159 -4.578 ...
#>  $ C07        : num  29.61 23.53 14.85 2.13 -7.79 ...
#>  $ C08        : num  34.1 21.1 11.63 2.94 -9.04 ...
#>  $ C09        : num  39.23 32.96 25.78 -6.18 -11.34 ...
#>  $ C10        : num  38.89 24.72 10.61 2.44 -8.79 ...
#>  $ C11        : num  47.18 24.02 11.48 3.26 1.75 ...
#>  $ C12        : num  101.9 79.1 29.7 11.9 -1.1 ...
#>  $ D01        : num  50.16 31.47 12.6 3.87 -5.23 ...
#>  $ D02        : num  2.17 0.164 10.127 15.747 3.482 ...
#>  $ D03        : num  67.953 54.416 20.895 -0.625 -17.136 ...
#>  $ D04        : num  53.72 33.32 19.88 1.23 -6.8 ...
#>  $ D05        : num  44.97 27.94 11.81 3.46 -11.11 ...
#>  $ D06        : num  31.79 26.09 13.25 3.06 -5.74 ...
#>  $ D07        : num  28.342 22.218 8.366 -1.349 -0.982 ...
#>  $ D08        : num  31.53 27.78 13.66 1.72 -6.16 ...
#>  $ D09        : num  28.13 21.54 18.02 1.8 -8.69 ...
#>  $ D10        : num  37.47 25.07 17.28 4.35 -6.84 ...
#>  $ D11        : num  -30.8 70.4 39 13.1 -7 ...
#>  $ D12        : num  47.66 39.5 26.61 14.68 3.43 ...
#>  $ E01        : num  40.543 22.803 8.124 0.362 -3.432 ...
#>  $ E02        : num  -25.84 -40.81 -6.94 36.48 22.84 ...
#>  $ E03        : num  36.38 30.77 23.15 15.09 9.75 ...
#>  $ E04        : num  62.877 37.715 16.614 -0.238 -11.835 ...
#>  $ E05        : num  36.97 24.2 10.26 3.29 -6.77 ...
#>  $ E06        : num  30.04 25.79 11.56 5.72 -10.29 ...
#>  $ E07        : num  34.325 25.109 15.401 0.344 -6.771 ...
#>  $ E08        : num  30.524 24.056 10.074 -0.442 -4.353 ...
#>  $ E09        : num  54.603 49.939 26.314 14.119 -0.906 ...
#>  $ E10        : num  36.05 28.3 15.2 3.06 -10.12 ...
#>  $ E11        : num  32.995 46.425 32.065 13.884 -0.258 ...
#>  $ E12        : num  27.94 1.19 -21.85 34.97 19.39 ...
#>  $ F01        : num  36.08 31.75 12.57 4.55 -8.34 ...
#>  $ F02        : num  40.36 24.7 12.19 1.82 -6.37 ...
#>  $ F03        : num  45.7 42.9 40.1 33.6 -34.3 ...
#>  $ F04        : num  23.8 -14.5 60.1 47.8 33.7 ...
#>  $ F05        : num  40.64 27.13 17.17 3.05 -6.52 ...
#>  $ F06        : num  38.77 38.03 28.57 13.59 -3.24 ...
#>  $ F07        : num  37.29 33.89 15.1 7.06 -7.44 ...
#>  $ F08        : num  35.87 23.53 19.89 2.93 -5.77 ...
#>  $ F09        : num  31.89 28.06 10.93 4.29 -6.12 ...
#>  $ F10        : num  47.48 38.67 24.59 6.09 -3.46 ...
#>  $ F11        : num  42.179 24.081 15.887 7.73 0.391 ...
#>  $ F12        : num  41.5 39.6 49.6 14.4 13.9 ...
#>  $ G01        : num  4.62 26.58 33.87 15.56 11.05 ...
#>  $ G02        : num  85.77 52.34 28.24 8.95 63.04 ...
#>  $ G03        : num  87.83 62.54 36.21 17.35 3.11 ...
#>  $ G04        : num  39.63 25.78 17.65 6.26 -1.28 ...
#>  $ G05        : num  38.5 14.1 -18.9 -49.3 -72 ...
#>  $ G06        : num  43.171 32.534 19.822 -0.755 -3.614 ...
#>  $ G07        : num  41.66 27.04 19.15 3.78 -5.6 ...
#>  $ G08        : num  48.38 33.39 15.53 3.9 -7.66 ...
#>  $ G09        : num  45 30.8 17.66 6.96 -9.24 ...
#>  $ G10        : num  98.2 80.6 59.1 26.4 -16 ...
#>  $ G11        : num  98.81 62.79 34.68 7.12 -21.11 ...
#>  $ G12        : num  -3.9 -5.99 -3.69 -3.96 3.18 ...
#>  $ H01        : num  97.56 87.47 21.41 -4.46 -25.88 ...
#>  $ H02        : num  75.83 56.78 36.55 32.43 2.59 ...
#>  $ H03        : num  94.52 64.68 38.37 21.72 -3.34 ...
#>  $ H04        : num  38.6 48.87 10.02 1.09 -11.95 ...
#>  $ H05        : num  55.54 27.892 16.687 0.986 -3.209 ...
#>  $ H06        : num  37.24 31.49 9.68 7.45 -9.08 ...
#>  $ H07        : num  41.05 27.83 14.17 2.5 -6.22 ...
#>  $ H08        : num  52.26 28.27 29.08 1.01 -6.97 ...
#>  $ H09        : num  45.39 37.98 17.15 8.14 -8.38 ...
#>  $ H10        : num  -240.4 -104.5 -22.1 39.9 57.5 ...
#>  $ H11        : num  77.8 49.9 35 13.8 31.2 ...
#>  $ H12        : num  -316.9 -118.1 -21.4 29.2 53.8 ...