Melting Curves (resolution: 0.2 degrees Celsius / step) for the Gene adk Using the Bio-Rad iQ5 thermo cycler

A melting curve for adk was performed.

data("C316.melt.hr")

Format

A data frame with 40 observations on the following 97 variables. The first column ("T") contains the temperature (resolution: 0.2 degrees Celsius / step) and consecutive columns contain the replicates ("A01" to "H12").

Details

adk was amplified in the Bio-Rad iQ5. After PCR, the temperature-dependent change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: C316.amp

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

data(C316.melt.hr)
str(C316.melt.hr)

#> 'data.frame':	201 obs. of  97 variables:
#>  $ T  : num  55 55.2 55.4 55.6 55.8 56 56.2 56.4 56.6 56.8 ...
#>  $ A01: num  52983 50495 48879 48186 47819 ...
#>  $ A02: num  46327 44323 43053 42553 42274 ...
#>  $ A03: num  47831 45846 44580 44080 43811 ...
#>  $ A04: num  46648 44490 43127 42570 42277 ...
#>  $ A05: num  50848 48655 47238 46657 46334 ...
#>  $ A06: num  52367 49940 48401 47814 47514 ...
#>  $ A07: num  58448 55991 54434 53792 53454 ...
#>  $ A08: num  57214 54773 53227 52622 52296 ...
#>  $ A09: num  51291 49037 47608 47064 46751 ...
#>  $ A10: num  47976 45939 44642 44133 43838 ...
#>  $ A11: num  53424 51119 49645 49050 48713 ...
#>  $ A12: num  53429 51230 49815 49224 48881 ...
#>  $ B01: num  48747 46631 45276 44720 44402 ...
#>  $ B02: num  51458 49086 47570 46967 46627 ...
#>  $ B03: num  40524 38668 37465 36968 36694 ...
#>  $ B04: num  54392 51799 50141 49486 49124 ...
#>  $ B05: num  46372 44202 42804 42266 41979 ...
#>  $ B06: num  41799 39795 38530 38048 37797 ...
#>  $ B07: num  51297 49057 47633 47074 46776 ...
#>  $ B08: num  48769 46727 45442 44940 44675 ...
#>  $ B09: num  51059 48964 47628 47126 46832 ...
#>  $ B10: num  53129 50886 49435 48843 48501 ...
#>  $ B11: num  49647 47425 45998 45399 45119 ...
#>  $ B12: num  46280 44266 43023 42571 42308 ...
#>  $ C01: num  61167 58486 56769 56075 55685 ...
#>  $ C02: num  45750 43731 42440 41918 41633 ...
#>  $ C03: num  46824 44801 43526 43033 42754 ...
#>  $ C04: num  50928 48583 47110 46534 46229 ...
#>  $ C05: num  46926 44924 43627 43117 42834 ...
#>  $ C06: num  45122 43102 41831 41342 41088 ...
#>  $ C07: num  49214 47103 45775 45256 44978 ...
#>  $ C08: num  43990 42099 40912 40476 40228 ...
#>  $ C09: num  51019 48983 47704 47227 46976 ...
#>  $ C10: num  48419 45855 44263 43705 43418 ...
#>  $ C11: num  47520 45656 44485 44034 43775 ...
#>  $ C12: num  50918 48671 47216 46577 46206 ...
#>  $ D01: num  48398 46223 44829 44256 43951 ...
#>  $ D02: num  47528 45147 43609 42991 42663 ...
#>  $ D03: num  43347 41367 40121 39631 39379 ...
#>  $ D04: num  49678 47311 45806 45206 44883 ...
#>  $ D05: num  48254 46093 44727 44201 43919 ...
#>  $ D06: num  47914 45726 44355 43832 43560 ...
#>  $ D07: num  47134 45230 44043 43595 43358 ...
#>  $ D08: num  51836 49556 48134 47591 47309 ...
#>  $ D09: num  49252 47149 45852 45373 45116 ...
#>  $ D10: num  51158 48923 47507 46951 46646 ...
#>  $ D11: num  49274 47270 45990 45487 45207 ...
#>  $ D12: num  54002 51676 50189 49555 49212 ...
#>  $ E01: num  51849 49461 47933 47316 46960 ...
#>  $ E02: num  44525 42483 41191 40671 40393 ...
#>  $ E03: num  44331 42495 41345 40888 40626 ...
#>  $ E04: num  49350 47011 45536 44959 44660 ...
#>  $ E05: num  44676 42718 41482 40994 40725 ...
#>  $ E06: num  47140 45302 44148 43705 43455 ...
#>  $ E07: num  46411 44606 43476 43055 42828 ...
#>  $ E08: num  48421 46355 45045 44541 44267 ...
#>  $ E09: num  48124 46172 44958 44500 44264 ...
#>  $ E10: num  54600 52243 50752 50186 49895 ...
#>  $ E11: num  47019 44760 43372 42880 42606 ...
#>  $ E12: num  42199 40547 39517 39132 38905 ...
#>  $ F01: num  53215 51048 49668 49110 48785 ...
#>  $ F02: num  38709 37292 36396 36032 35821 ...
#>  $ F03: num  48956 47130 45976 45514 45238 ...
#>  $ F04: num  47721 45887 44729 44256 43985 ...
#>  $ F05: num  49412 47747 46697 46267 46016 ...
#>  $ F06: num  48329 46761 45774 45391 45150 ...
#>  $ F07: num  46985 45520 44607 44272 44044 ...
#>  $ F08: num  50802 49116 48071 47687 47448 ...
#>  $ F09: num  48356 46685 45652 45265 45031 ...
#>  $ F10: num  44647 43055 42053 41661 41424 ...
#>  $ F11: num  48486 46556 45362 44908 44650 ...
#>  $ F12: num  44443 42748 41662 41217 40953 ...
#>  $ G01: num  55440 53334 51944 51340 50982 ...
#>  $ G02: num  66369 63858 62243 61578 61158 ...
#>  $ G03: num  60575 58428 57052 56489 56139 ...
#>  $ G04: num  57746 55750 54476 53943 53615 ...
#>  $ G05: num  47063 45595 44652 44281 44036 ...
#>  $ G06: num  58429 56607 55436 54956 54655 ...
#>  $ G07: num  61983 60112 58919 58451 58138 ...
#>  $ G08: num  67130 65238 64046 63561 63227 ...
#>  $ G09: num  63118 61113 59863 59388 59102 ...
#>  $ G10: num  62212 60065 58681 58060 57764 ...
#>  $ G11: num  65794 63853 62619 62122 61778 ...
#>  $ G12: num  57819 55717 54443 53996 53693 ...
#>  $ H01: num  64549 62667 61462 60970 60655 ...
#>  $ H02: num  44825 43448 42567 42190 41964 ...
#>  $ H03: num  58077 56386 55289 54819 54502 ...
#>  $ H04: num  51824 50220 49193 48769 48488 ...
#>  $ H05: num  59726 57975 56846 56387 56063 ...
#>  $ H06: num  63930 62182 61052 60578 60257 ...
#>  $ H07: num  61325 59872 58942 58549 58260 ...
#>  $ H08: num  66374 64555 63403 62962 62631 ...
#>  $ H09: num  64474 62818 61771 61338 61035 ...
#>  $ H10: num  62154 60855 60031 59656 59385 ...
#>  $ H11: num  57692 56302 55411 55033 54754 ...
#>  $ H12: num  65335 63564 62328 61636 61314 ...