Melting Curves (resolution: 0.5 degrees Celsius / step) for the Gene adk Using the Bio-Rad iQ5 thermo cycler

A melting curve for adk was performed.

data("C316.melt.lr")

Format

A data frame with 40 observations on the following 97 variables. The first column ("T") contains the temperature (resolution: 0.5 degrees Celsius / step) and consecutive columns contain the replicates ("A01" to "H12").

Details

adk was amplified in the Bio-Rad iQ5. After PCR, the temperature-dependent change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: C316.amp

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

data(C316.melt.lr)
str(C316.melt.lr)

#> 'data.frame':	81 obs. of  97 variables:
#>  $ T  : num  55 55.5 56 56.5 57 57.5 58 58.5 59 59.5 ...
#>  $ A01: num  41947 44670 46167 46286 45841 ...
#>  $ A02: num  37901 40136 41327 41372 40967 ...
#>  $ A03: num  39190 41493 42722 42784 42358 ...
#>  $ A04: num  37493 40074 41519 41697 41325 ...
#>  $ A05: num  41349 43970 45406 45528 45108 ...
#>  $ A06: num  42342 45143 46692 46887 46463 ...
#>  $ A07: num  47389 50296 51909 52046 51617 ...
#>  $ A08: num  46859 49849 51491 51617 51143 ...
#>  $ A09: num  42413 45175 46670 46803 46337 ...
#>  $ A10: num  40330 42724 44013 44089 43648 ...
#>  $ A11: num  44334 47014 48428 48457 47942 ...
#>  $ A12: num  45190 47673 48932 48866 48324 ...
#>  $ B01: num  39789 42189 43460 43492 43053 ...
#>  $ B02: num  41233 43926 45388 45547 45113 ...
#>  $ B03: num  32846 34983 36175 36310 35994 ...
#>  $ B04: num  43785 46749 48413 48627 48200 ...
#>  $ B05: num  37031 39611 41054 41249 40872 ...
#>  $ B06: num  33529 35816 37110 37310 36976 ...
#>  $ B07: num  41764 44473 45972 46124 45720 ...
#>  $ B08: num  40176 42707 44094 44224 43797 ...
#>  $ B09: num  42964 45589 46986 47065 46603 ...
#>  $ B10: num  45063 47812 49286 49344 48844 ...
#>  $ B11: num  42531 44945 46217 46227 45734 ...
#>  $ B12: num  40069 42184 43235 43160 42659 ...
#>  $ C01: num  49929 52913 54486 54555 54004 ...
#>  $ C02: num  37123 39465 40746 40851 40456 ...
#>  $ C03: num  38457 40808 42054 42106 41684 ...
#>  $ C04: num  41305 44038 45556 45732 45331 ...
#>  $ C05: num  38551 40930 42237 42344 41947 ...
#>  $ C06: num  36555 39033 40420 40611 40268 ...
#>  $ C07: num  40455 43086 44553 44711 44327 ...
#>  $ C08: num  36531 38868 40158 40292 39925 ...
#>  $ C09: num  43070 45650 47035 47111 46635 ...
#>  $ C10: num  39184 41986 43558 43763 43378 ...
#>  $ C11: num  40261 43033 44480 44495 44032 ...
#>  $ C12: num  42081 44677 46074 46154 45697 ...
#>  $ D01: num  39258 41747 43118 43239 42846 ...
#>  $ D02: num  37051 39764 41294 41528 41177 ...
#>  $ D03: num  34975 37216 38426 38525 38162 ...
#>  $ D04: num  39601 42451 44055 44285 43908 ...
#>  $ D05: num  39377 42065 43574 43768 43389 ...
#>  $ D06: num  38720 41451 42999 43234 42870 ...
#>  $ D07: num  38963 41407 42747 42880 42481 ...
#>  $ D08: num  42817 45707 47326 47519 47104 ...
#>  $ D09: num  41270 43792 45135 45209 44750 ...
#>  $ D10: num  42626 45446 47003 47139 46687 ...
#>  $ D11: num  41562 44079 45412 45432 44962 ...
#>  $ D12: num  44404 47622 49318 49414 48902 ...
#>  $ E01: num  41779 44539 46016 46144 45688 ...
#>  $ E02: num  35779 38155 39451 39584 39218 ...
#>  $ E03: num  36317 38502 39652 39691 39283 ...
#>  $ E04: num  39255 42068 43664 43901 43529 ...
#>  $ E05: num  36611 39032 40381 40518 40155 ...
#>  $ E06: num  39722 42156 43457 43538 43125 ...
#>  $ E07: num  38962 41236 42450 42541 42118 ...
#>  $ E08: num  39879 42549 44015 44158 43762 ...
#>  $ E09: num  40311 42781 44091 44163 43742 ...
#>  $ E10: num  45388 48288 49864 49983 49508 ...
#>  $ E11: num  38926 41567 43023 43250 42651 ...
#>  $ E12: num  35876 37884 38907 38873 38446 ...
#>  $ F01: num  44450 46920 48190 48169 47668 ...
#>  $ F02: num  32933 34631 35484 35425 35025 ...
#>  $ F03: num  41678 43861 44957 44892 44386 ...
#>  $ F04: num  39785 42000 43160 43162 42713 ...
#>  $ F05: num  42412 44597 45700 45629 45138 ...
#>  $ F06: num  42167 44287 45358 45315 44826 ...
#>  $ F07: num  40666 42658 43621 43542 43040 ...
#>  $ F08: num  44273 46553 47706 47626 47110 ...
#>  $ F09: num  41456 43719 44870 44847 44364 ...
#>  $ F10: num  38270 40284 41282 41207 40733 ...
#>  $ F11: num  41284 43547 44705 44667 44183 ...
#>  $ F12: num  38413 40409 41400 41311 40831 ...
#>  $ G01: num  46512 48877 50058 49967 49418 ...
#>  $ G02: num  55306 58205 59632 59507 58819 ...
#>  $ G03: num  51436 54033 55353 55279 54667 ...
#>  $ G04: num  49035 51466 52676 52545 51948 ...
#>  $ G05: num  41015 43006 43980 43860 43371 ...
#>  $ G06: num  50761 53231 54439 54348 53739 ...
#>  $ G07: num  53595 56153 57411 57307 56664 ...
#>  $ G08: num  59060 61724 62989 62781 62037 ...
#>  $ G09: num  55493 58143 59435 59298 58655 ...
#>  $ G10: num  54159 56626 57779 57556 56864 ...
#>  $ G11: num  58840 61323 62453 62126 61365 ...
#>  $ G12: num  51293 53372 54292 53990 53305 ...
#>  $ H01: num  56593 58924 59972 59643 58882 ...
#>  $ H02: num  38703 40404 41216 41054 40589 ...
#>  $ H03: num  50278 52384 53353 53115 52464 ...
#>  $ H04: num  44340 46340 47307 47160 46620 ...
#>  $ H05: num  51029 53304 54383 54220 53597 ...
#>  $ H06: num  55749 58046 59133 58906 58242 ...
#>  $ H07: num  54554 56577 57491 57192 56518 ...
#>  $ H08: num  58302 60612 61665 61396 60688 ...
#>  $ H09: num  57466 59742 60735 60423 59684 ...
#>  $ H10: num  57528 59253 59934 59501 58748 ...
#>  $ H11: num  53036 54842 55574 55168 54459 ...
#>  $ H12: num  60059 62303 63208 62715 61787 ...