qPCR Experiment for the Amplification of adk Using the Bio-Rad CFX96 thermo cycler

A quantitative real-time PCR of adk was performed.

data("C317.amp")

Format

A data frame with 40 observations on the following 97 variables. The first column ("Cycle") contains the number of cycles and consecutive columns contain the replicates ("A01" to "H12").

Details

adk was amplified in the Bio-Rad CFX96. The change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

gDNA: 28.43 ng/microL DNA concentration, 260/280 ratio= 1.96

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: 10 minutes at 95 degrees Celsius 40 x 30 seconds at 95 degrees Celsius 45 seconds at 59 degrees Celsius 45 seconds at 68 degrees Celsius

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

  data(C317.amp)
  str(C317.amp)

#> 'data.frame':	40 obs. of  97 variables:
#>  $ Cycle: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ A1   : num  2330 2343 2355 2359 2364 ...
#>  $ A2   : num  2361 2367 2370 2381 2375 ...
#>  $ A3   : num  2364 2373 2377 2385 2391 ...
#>  $ A4   : num  2347 2355 2362 2362 2367 ...
#>  $ A5   : num  2336 2356 2361 2369 2371 ...
#>  $ A6   : num  2313 2322 2331 2332 2338 ...
#>  $ A7   : num  2345 2358 2363 2371 2371 ...
#>  $ A8   : num  2356 2367 2377 2384 2379 ...
#>  $ A9   : num  2352 2362 2368 2374 2371 ...
#>  $ A10  : num  2350 2360 2359 2366 2360 ...
#>  $ A11  : num  2353 2369 2376 2372 2371 ...
#>  $ A12  : num  2207 2212 2218 2219 2224 ...
#>  $ B1   : num  2350 2358 2365 2367 2369 ...
#>  $ B2   : num  2376 2381 2386 2386 2393 ...
#>  $ B3   : num  2368 2373 2375 2378 2377 ...
#>  $ B4   : num  2369 2375 2381 2381 2379 ...
#>  $ B5   : num  2357 2362 2369 2372 2372 ...
#>  $ B6   : num  2393 2401 2396 2406 2406 ...
#>  $ B7   : num  2382 2387 2394 2396 2393 ...
#>  $ B8   : num  2395 2404 2408 2412 2407 ...
#>  $ B9   : num  2382 2394 2395 2399 2399 ...
#>  $ B10  : num  2379 2386 2384 2386 2388 ...
#>  $ B11  : num  2427 2436 2439 2443 2441 ...
#>  $ B12  : num  2341 2352 2356 2360 2358 ...
#>  $ C1   : num  2352 2360 2371 2376 2380 ...
#>  $ C2   : num  2378 2386 2387 2392 2389 ...
#>  $ C3   : num  2378 2387 2388 2386 2386 ...
#>  $ C4   : num  2400 2409 2409 2408 2407 ...
#>  $ C5   : num  2408 2412 2419 2422 2421 ...
#>  $ C6   : num  2439 2447 2450 2454 2451 ...
#>  $ C7   : num  2411 2421 2421 2425 2423 ...
#>  $ C8   : num  2426 2435 2432 2441 2444 ...
#>  $ C9   : num  2366 2369 2373 2379 2373 ...
#>  $ C10  : num  2374 2377 2382 2385 2385 ...
#>  $ C11  : num  2405 2407 2409 2415 2415 ...
#>  $ C12  : num  2362 2373 2376 2384 2382 ...
#>  $ D1   : num  2358 2362 2371 2369 2377 ...
#>  $ D2   : num  2367 2376 2386 2387 2390 ...
#>  $ D3   : num  2395 2406 2410 2415 2415 ...
#>  $ D4   : num  2437 2442 2447 2457 2452 ...
#>  $ D5   : num  2445 2455 2454 2461 2459 ...
#>  $ D6   : num  2451 2459 2465 2467 2469 ...
#>  $ D7   : num  2441 2445 2452 2455 2451 ...
#>  $ D8   : num  2421 2432 2434 2436 2433 ...
#>  $ D9   : num  2400 2413 2409 2422 2419 ...
#>  $ D10  : num  2437 2443 2440 2450 2445 ...
#>  $ D11  : num  2439 2444 2446 2451 2450 ...
#>  $ D12  : num  2411 2413 2409 2413 2412 ...
#>  $ E1   : num  2302 2309 2310 2316 2319 ...
#>  $ E2   : num  2388 2396 2397 2397 2401 ...
#>  $ E3   : num  2423 2428 2436 2438 2439 ...
#>  $ E4   : num  2412 2427 2429 2437 2438 ...
#>  $ E5   : num  2426 2438 2440 2442 2440 ...
#>  $ E6   : num  2428 2430 2436 2438 2437 ...
#>  $ E7   : num  2463 2469 2471 2476 2471 ...
#>  $ E8   : num  2781 2806 2791 2787 2776 ...
#>  $ E9   : num  2403 2408 2411 2418 2416 ...
#>  $ E10  : num  2417 2420 2425 2428 2427 ...
#>  $ E11  : num  2423 2428 2430 2433 2434 ...
#>  $ E12  : num  2398 2402 2403 2405 2406 ...
#>  $ F1   : num  2312 2318 2324 2331 2336 ...
#>  $ F2   : num  2367 2375 2379 2381 2385 ...
#>  $ F3   : num  2373 2384 2387 2387 2388 ...
#>  $ F4   : num  2419 2426 2427 2433 2436 ...
#>  $ F5   : num  2403 2415 2413 2421 2421 ...
#>  $ F6   : num  2415 2424 2425 2431 2430 ...
#>  $ F7   : num  2442 2445 2447 2450 2454 ...
#>  $ F8   : num  2423 2427 2426 2429 2427 ...
#>  $ F9   : num  2408 2412 2416 2421 2413 ...
#>  $ F10  : num  2423 2426 2426 2431 2431 ...
#>  $ F11  : num  2433 2438 2434 2436 2439 ...
#>  $ F12  : num  2447 2455 2456 2457 2456 ...
#>  $ G1   : num  2256 2262 2266 2270 2269 ...
#>  $ G2   : num  2346 2352 2356 2361 2365 ...
#>  $ G3   : num  2328 2338 2336 2343 2344 ...
#>  $ G4   : num  2335 2340 2342 2346 2344 ...
#>  $ G5   : num  2343 2353 2351 2357 2355 ...
#>  $ G6   : num  2336 2348 2346 2349 2351 ...
#>  $ G7   : num  2328 2338 2336 2347 2350 ...
#>  $ G8   : num  2353 2361 2358 2366 2364 ...
#>  $ G9   : num  2343 2350 2354 2353 2360 ...
#>  $ G10  : num  2351 2358 2361 2362 2364 ...
#>  $ G11  : num  2368 2378 2379 2390 2380 ...
#>  $ G12  : num  2362 2373 2380 2382 2387 ...
#>  $ H1   : num  2258 2257 2262 2267 2270 ...
#>  $ H2   : num  2345 2360 2358 2363 2367 ...
#>  $ H3   : num  2334 2345 2350 2351 2358 ...
#>  $ H4   : num  2354 2359 2364 2366 2367 ...
#>  $ H5   : num  2329 2335 2341 2350 2349 ...
#>  $ H6   : num  2336 2348 2347 2350 2351 ...
#>  $ H7   : num  2337 2351 2356 2352 2357 ...
#>  $ H8   : num  2362 2363 2377 2378 2384 ...
#>  $ H9   : num  2360 2368 2369 2375 2376 ...
#>  $ H10  : num  2329 2347 2342 2350 2351 ...
#>  $ H11  : num  2365 2377 2375 2383 2376 ...
#>  $ H12  : num  2353 2361 2379 2381 2377 ...