Melting Curves (resolution: 0.1 degrees Celsius / step) for the Gene adk Using the Bio-Rad CFX96 thermo cycler

A melting curve for adk was performed.

data("C317.melt.hr")

Format

A data frame with 40 observations on the following 97 variables. The first column ("Temperature") contains the temperature (resolution: 0.1 degrees Celsius / step) and consecutive columns contain the replicates ("A1" to "H12").

Details

adk was amplified in the Bio-Rad CFX96. After PCR, the temperature-dependent change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: C317.amp

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

  data(C317.melt.hr)
  str(C317.melt.hr)

#> 'data.frame':	451 obs. of  97 variables:
#>  $ Temperature: num  50 50.1 50.2 50.3 50.4 50.5 50.6 50.7 50.8 50.9 ...
#>  $ A1         : num  6209 6192 6176 6159 6143 ...
#>  $ A2         : num  6802 6784 6765 6747 6728 ...
#>  $ A3         : num  6848 6830 6812 6794 6776 ...
#>  $ A4         : num  6595 6577 6560 6542 6525 ...
#>  $ A5         : num  6624 6604 6584 6564 6545 ...
#>  $ A6         : num  6163 6147 6132 6116 6101 ...
#>  $ A7         : num  6658 6643 6627 6612 6596 ...
#>  $ A8         : num  6632 6614 6597 6579 6562 ...
#>  $ A9         : num  6935 6917 6899 6881 6863 ...
#>  $ A10        : num  6784 6768 6752 6735 6719 ...
#>  $ A11        : num  7002 6987 6971 6956 6941 ...
#>  $ A12        : num  4655 4644 4632 4621 4610 ...
#>  $ B1         : num  6437 6420 6403 6387 6370 ...
#>  $ B2         : num  6677 6660 6644 6627 6610 ...
#>  $ B3         : num  6813 6795 6776 6758 6739 ...
#>  $ B4         : num  6971 6954 6937 6919 6902 ...
#>  $ B5         : num  6815 6799 6782 6766 6749 ...
#>  $ B6         : num  7098 7082 7065 7049 7033 ...
#>  $ B7         : num  6979 6963 6947 6932 6916 ...
#>  $ B8         : num  7151 7136 7122 7107 7092 ...
#>  $ B9         : num  7194 7179 7164 7150 7135 ...
#>  $ B10        : num  6939 6928 6916 6904 6892 ...
#>  $ B11        : num  7406 7393 7380 7367 7354 ...
#>  $ B12        : num  6708 6697 6686 6675 6664 ...
#>  $ C1         : num  6895 6879 6863 6847 6831 ...
#>  $ C2         : num  6912 6896 6880 6864 6848 ...
#>  $ C3         : num  6747 6730 6713 6696 6679 ...
#>  $ C4         : num  7043 7026 7009 6992 6975 ...
#>  $ C5         : num  7160 7142 7124 7106 7088 ...
#>  $ C6         : num  7405 7387 7370 7353 7336 ...
#>  $ C7         : num  7254 7237 7221 7205 7188 ...
#>  $ C8         : num  7132 7118 7103 7089 7074 ...
#>  $ C9         : num  7168 7154 7140 7127 7113 ...
#>  $ C10        : num  7307 7294 7281 7268 7255 ...
#>  $ C11        : num  7446 7431 7416 7400 7385 ...
#>  $ C12        : num  7059 7044 7029 7014 6999 ...
#>  $ D1         : num  6519 6503 6486 6470 6454 ...
#>  $ D2         : num  7075 7056 7038 7020 7001 ...
#>  $ D3         : num  7399 7380 7361 7342 7323 ...
#>  $ D4         : num  7347 7328 7308 7289 7269 ...
#>  $ D5         : num  6912 6895 6878 6860 6843 ...
#>  $ D6         : num  7666 7651 7635 7619 7603 ...
#>  $ D7         : num  7613 7597 7582 7566 7551 ...
#>  $ D8         : num  7004 6991 6978 6965 6952 ...
#>  $ D9         : num  7261 7247 7232 7218 7204 ...
#>  $ D10        : num  7910 7897 7885 7872 7859 ...
#>  $ D11        : num  7839 7829 7818 7807 7797 ...
#>  $ D12        : num  7507 7495 7484 7473 7461 ...
#>  $ E1         : num  6217 6202 6186 6170 6154 ...
#>  $ E2         : num  7333 7317 7300 7283 7267 ...
#>  $ E3         : num  7300 7283 7266 7249 7232 ...
#>  $ E4         : num  7046 7029 7013 6996 6980 ...
#>  $ E5         : num  7389 7372 7355 7337 7320 ...
#>  $ E6         : num  7779 7761 7744 7726 7709 ...
#>  $ E7         : num  7512 7494 7477 7459 7442 ...
#>  $ E8         : num  7766 7750 7734 7717 7701 ...
#>  $ E9         : num  7664 7649 7634 7619 7604 ...
#>  $ E10        : num  7742 7728 7714 7700 7686 ...
#>  $ E11        : num  7880 7866 7853 7839 7826 ...
#>  $ E12        : num  7526 7510 7495 7480 7465 ...
#>  $ F1         : num  6151 6134 6116 6099 6081 ...
#>  $ F2         : num  6732 6716 6699 6682 6665 ...
#>  $ F3         : num  6967 6948 6930 6912 6894 ...
#>  $ F4         : num  7251 7233 7215 7197 7179 ...
#>  $ F5         : num  7162 7145 7127 7110 7092 ...
#>  $ F6         : num  7220 7201 7183 7165 7147 ...
#>  $ F7         : num  6939 6923 6907 6891 6874 ...
#>  $ F8         : num  7316 7299 7281 7264 7247 ...
#>  $ F9         : num  7459 7442 7425 7408 7391 ...
#>  $ F10        : num  7478 7464 7451 7438 7425 ...
#>  $ F11        : num  7361 7347 7332 7318 7304 ...
#>  $ F12        : num  7727 7713 7699 7685 7671 ...
#>  $ G1         : num  6017 6001 5984 5968 5952 ...
#>  $ G2         : num  6718 6701 6685 6668 6652 ...
#>  $ G3         : num  6639 6623 6607 6591 6574 ...
#>  $ G4         : num  6851 6834 6818 6802 6785 ...
#>  $ G5         : num  7020 7002 6985 6967 6950 ...
#>  $ G6         : num  6914 6897 6880 6864 6847 ...
#>  $ G7         : num  6833 6819 6805 6790 6776 ...
#>  $ G8         : num  7132 7115 7098 7080 7063 ...
#>  $ G9         : num  7245 7231 7217 7203 7189 ...
#>  $ G10        : num  7126 7112 7098 7085 7071 ...
#>  $ G11        : num  7516 7499 7483 7467 7451 ...
#>  $ G12        : num  7096 7081 7065 7050 7034 ...
#>  $ H1         : num  5676 5659 5643 5627 5610 ...
#>  $ H2         : num  6644 6627 6610 6593 6576 ...
#>  $ H3         : num  6637 6621 6604 6587 6571 ...
#>  $ H4         : num  6838 6822 6805 6789 6773 ...
#>  $ H5         : num  6773 6756 6740 6723 6706 ...
#>  $ H6         : num  6787 6769 6751 6734 6716 ...
#>  $ H7         : num  6984 6967 6951 6935 6919 ...
#>  $ H8         : num  7088 7073 7058 7042 7027 ...
#>  $ H9         : num  7338 7324 7310 7296 7282 ...
#>  $ H10        : num  7282 7269 7256 7243 7230 ...
#>  $ H11        : num  7539 7524 7509 7495 7480 ...
#>  $ H12        : num  7529 7514 7499 7485 7470 ...