Melting Curves (resolution: 0.5 degrees Celsius / step) for the Gene adk Using the Bio-Rad CFX96 thermo cycler

A melting curve for adk was performed.

data("C317.melt.lr")

Format

A data frame with 40 observations on the following 97 variables. The first column ("Temperature") contains the temperature (resolution: 0.5 degrees Celsius / step) and consecutive columns contain the replicates ("A1" to "H12").

Details

adk was amplified in the Bio-Rad CFX96. After PCR, the temperature-dependent change of fluorescence was simultaneously monitored (EvaGreen, Mao et al. 2007). The primer sequences for adk were taken from this study.

adk fw: CTCAGGCTCAGTTCATCATGGA adk rv: AGTTTGCCAGCATCCATAATGTC

PCR conditions: C317.amp

Source

Stefan Roediger, Claudia Deutschmann, Claudia Zelck (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

  data(C317.melt.lr)
  str(C317.melt.lr)

#> 'data.frame':	91 obs. of  97 variables:
#>  $ Temperature: num  50 50.5 51 51.5 52 52.5 53 53.5 54 54.5 ...
#>  $ A1         : num  6438 6385 6332 6278 6225 ...
#>  $ A2         : num  7080 7018 6956 6894 6832 ...
#>  $ A3         : num  7189 7125 7061 6998 6934 ...
#>  $ A4         : num  6972 6909 6846 6784 6721 ...
#>  $ A5         : num  6976 6913 6850 6787 6724 ...
#>  $ A6         : num  6614 6552 6490 6428 6366 ...
#>  $ A7         : num  7095 7031 6967 6902 6838 ...
#>  $ A8         : num  6934 6873 6812 6751 6689 ...
#>  $ A9         : num  7349 7283 7216 7149 7082 ...
#>  $ A10        : num  7085 7023 6961 6899 6837 ...
#>  $ A11        : num  7397 7333 7268 7203 7139 ...
#>  $ A12        : num  4996 4952 4908 4865 4821 ...
#>  $ B1         : num  6770 6712 6653 6594 6536 ...
#>  $ B2         : num  7098 7035 6973 6910 6848 ...
#>  $ B3         : num  7222 7158 7094 7030 6966 ...
#>  $ B4         : num  7448 7380 7313 7245 7178 ...
#>  $ B5         : num  7338 7271 7205 7138 7072 ...
#>  $ B6         : num  7645 7575 7504 7434 7363 ...
#>  $ B7         : num  7533 7465 7396 7328 7260 ...
#>  $ B8         : num  7662 7592 7522 7452 7382 ...
#>  $ B9         : num  7722 7651 7579 7508 7436 ...
#>  $ B10        : num  7467 7398 7330 7261 7192 ...
#>  $ B11        : num  7911 7838 7766 7693 7621 ...
#>  $ B12        : num  7100 7037 6974 6911 6849 ...
#>  $ C1         : num  7260 7196 7133 7069 7006 ...
#>  $ C2         : num  7393 7327 7261 7195 7129 ...
#>  $ C3         : num  7212 7147 7082 7017 6952 ...
#>  $ C4         : num  7598 7528 7459 7389 7319 ...
#>  $ C5         : num  7798 7724 7651 7578 7505 ...
#>  $ C6         : num  7900 7828 7757 7685 7613 ...
#>  $ C7         : num  7815 7744 7672 7600 7528 ...
#>  $ C8         : num  7729 7657 7586 7514 7442 ...
#>  $ C9         : num  7711 7640 7570 7499 7428 ...
#>  $ C10        : num  7884 7811 7739 7666 7593 ...
#>  $ C11        : num  7976 7902 7828 7755 7681 ...
#>  $ C12        : num  7464 7396 7328 7260 7192 ...
#>  $ D1         : num  6899 6838 6778 6717 6657 ...
#>  $ D2         : num  7532 7464 7395 7327 7258 ...
#>  $ D3         : num  7908 7837 7766 7695 7624 ...
#>  $ D4         : num  7806 7736 7666 7596 7527 ...
#>  $ D5         : num  7499 7432 7365 7298 7230 ...
#>  $ D6         : num  8185 8111 8036 7962 7888 ...
#>  $ D7         : num  8233 8158 8084 8010 7936 ...
#>  $ D8         : num  7561 7494 7427 7360 7292 ...
#>  $ D9         : num  7885 7812 7739 7666 7593 ...
#>  $ D10        : num  8479 8400 8321 8242 8163 ...
#>  $ D11        : num  8483 8404 8325 8247 8168 ...
#>  $ D12        : num  7957 7885 7814 7742 7670 ...
#>  $ E1         : num  6605 6546 6488 6429 6371 ...
#>  $ E2         : num  7850 7780 7710 7639 7569 ...
#>  $ E3         : num  7805 7735 7665 7595 7525 ...
#>  $ E4         : num  7519 7450 7382 7314 7245 ...
#>  $ E5         : num  7907 7833 7759 7685 7611 ...
#>  $ E6         : num  8200 8124 8049 7974 7898 ...
#>  $ E7         : num  8085 8011 7937 7863 7789 ...
#>  $ E8         : num  8321 8244 8167 8091 8014 ...
#>  $ E9         : num  8358 8278 8198 8118 8038 ...
#>  $ E10        : num  8347 8269 8191 8113 8035 ...
#>  $ E11        : num  8444 8366 8288 8210 8131 ...
#>  $ E12        : num  8107 8033 7959 7886 7812 ...
#>  $ F1         : num  6444 6386 6328 6270 6212 ...
#>  $ F2         : num  7189 7125 7061 6997 6933 ...
#>  $ F3         : num  7464 7395 7326 7257 7188 ...
#>  $ F4         : num  7734 7662 7591 7519 7448 ...
#>  $ F5         : num  7649 7578 7508 7438 7367 ...
#>  $ F6         : num  7642 7571 7499 7428 7356 ...
#>  $ F7         : num  7380 7311 7242 7174 7105 ...
#>  $ F8         : num  7750 7675 7601 7526 7451 ...
#>  $ F9         : num  7976 7899 7822 7744 7667 ...
#>  $ F10        : num  8121 8043 7964 7886 7807 ...
#>  $ F11        : num  8054 7977 7901 7824 7747 ...
#>  $ F12        : num  8325 8247 8169 8092 8014 ...
#>  $ G1         : num  6324 6266 6208 6150 6093 ...
#>  $ G2         : num  7127 7063 6999 6936 6872 ...
#>  $ G3         : num  7099 7034 6969 6904 6840 ...
#>  $ G4         : num  7263 7197 7131 7065 6999 ...
#>  $ G5         : num  7383 7315 7247 7178 7110 ...
#>  $ G6         : num  7369 7300 7232 7163 7095 ...
#>  $ G7         : num  7391 7323 7255 7188 7120 ...
#>  $ G8         : num  7574 7503 7431 7360 7288 ...
#>  $ G9         : num  7755 7682 7609 7536 7463 ...
#>  $ G10        : num  7692 7619 7546 7473 7401 ...
#>  $ G11        : num  8146 8068 7991 7913 7836 ...
#>  $ G12        : num  7502 7435 7369 7302 7235 ...
#>  $ H1         : num  5902 5852 5803 5754 5705 ...
#>  $ H2         : num  6949 6888 6827 6765 6704 ...
#>  $ H3         : num  6934 6873 6813 6753 6693 ...
#>  $ H4         : num  7190 7126 7062 6998 6934 ...
#>  $ H5         : num  7121 7057 6992 6928 6864 ...
#>  $ H6         : num  7089 7024 6960 6895 6830 ...
#>  $ H7         : num  7353 7286 7220 7153 7086 ...
#>  $ H8         : num  7582 7511 7440 7370 7299 ...
#>  $ H9         : num  7812 7740 7669 7597 7525 ...
#>  $ H10        : num  7723 7651 7579 7507 7435 ...
#>  $ H11        : num  8012 7937 7861 7786 7710 ...
#>  $ H12        : num  7883 7813 7743 7672 7602 ...