qPCR Experiment for the Amplification of MLC-2v and Vimentin (as decadic dilutions) Using the Roche Light Cycler 1.5

Dilution experiment and metling curve (C60.melt) for the human genes MLC-2v and Vimentin (see Roediger et al. 2013) using the Roch Light Cycler 1.5.

data(C60.amp)

Format

A data frame with 45 observations on the following 33 variables.

Index

Index of Cycles

Vim.0.1

Vimentin water control

Vim.0.2

Vimentin water control

Vim.1.1

Vimentin 10^-3 diluted

Vim.1.2

Vimentin 10^-3 diluted

Vim.2.1

Vimentin 10^-4 diluted

Vim.2.2

Vimentin 10^-4 diluted

Vim.3.1

Vimentin 10^-5 diluted

Vim.3.2

Vimentin 10^-5 diluted

Vim.4.1

Vimentin 10^-6 diluted

Vim.4.2

Vimentin 10^-6 diluted

Vim.5.1

Vimentin 10^-7 diluted

Vim.5.2

Vimentin 10^-7 diluted

Vim.6.1

Vimentin 10^-8 diluted

Vim.6.2

Vimentin 10^-8 diluted

Vim.7.1

Vimentin 10^-9 diluted

Vim.7.2

Vimentin 10^-9 diluted

MLC2v.1.1

MLC-2v 10^-3 diluted

MLC2v.1.2

MLC-2v 10^-3 diluted

MLC2v.2.1

MLC-2v 10^-4 diluted

MLC2v.2.2

MLC-2v 10^-4 diluted

MLC2v.3.1

MLC-2v 10^-5 diluted

MLC2v.3.2

MLC-2v 10^-5 diluted

MLC2v.4.1

MLC-2v 10^-6 diluted

MLC2v.4.2

MLC-2v 10^-6 diluted

MLC2v.5.1

MLC-2v 10^-7 diluted

MLC2v.5.2

MLC-2v 10^-7 diluted

MLC2v.6.1

MLC-2v 10^-8 diluted

MLC2v.6.2

MLC-2v 10^-8 diluted

MLC2v.7.1

MLC-2v 10^-9 diluted

MLC2v.7.2

MLC-2v 10^-9 diluted

MLC2v.0.1

MLC-2v water control

MLC2v.0.2

MLC-2v water control

Details

MLC-2v and Vimentin were amplified in the Roche Light Cycler 1.5. Decadic dilutions of the input cDNA were prepared. The change of fluorescence was simultaneously monitored with EvaGreen. The primer sequences for MLC-2v were taken from Roediger et al. (2013). A 10 micro L qPCR reaction was composed of 250 nM primer (forward and reverse), Roche qPCR Master-Mix (according to the manufactures recommendations) and 1 micro L input DNA. EvaGreen was used at 1x final. During the amplification was monitored 58 degrees Celsius. Temperature profile:

95 deg C for 8 minutes 40 x 95 deg C for 10 sec 58 deg C for 15 sec 69 deg C for 25 sec

Source

Stefan Roediger, Claudia Deutschmann (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

## data(C60.amp)
str(C60.amp)

#> 'data.frame':	45 obs. of  33 variables:
#>  $ Index    : int  0 1 2 3 4 5 6 7 8 9 ...
#>  $ Vim.0.1  : num  0.003967 0.000916 0.000916 -0.00061 -0.00061 ...
#>  $ Vim.0.2  : num  0.001221 0.001221 -0.000305 -0.000305 -0.000305 ...
#>  $ Vim.1.1  : num  -0.02716 -0.02564 -0.02106 -0.0119 0.00946 ...
#>  $ Vim.1.2  : num  -0.02869 -0.02716 -0.02258 -0.0119 0.00946 ...
#>  $ Vim.2.1  : num  0.00061 -0.000916 -0.000916 -0.000916 -0.000916 ...
#>  $ Vim.2.2  : num  0.00153 0 -0.00153 -0.00153 0 ...
#>  $ Vim.3.1  : num  0.001221 0.001221 -0.000305 -0.000305 -0.000305 ...
#>  $ Vim.3.2  : num  0.002136 0.00061 0.00061 -0.000916 0.00061 ...
#>  $ Vim.4.1  : num  0.003662 0.003662 0.00061 -0.000916 -0.000916 ...
#>  $ Vim.4.2  : num  0.00305 0.00305 0 0 -0.00153 ...
#>  $ Vim.5.1  : num  0.001221 0.001221 -0.000305 -0.000305 -0.000305 ...
#>  $ Vim.5.2  : num  0.000305 0.001831 0.000305 0.000305 -0.001221 ...
#>  $ Vim.6.1  : num  0.001221 0.001221 -0.000305 -0.000305 -0.000305 ...
#>  $ Vim.6.2  : num  0.002136 0.002136 0.00061 -0.000916 -0.000916 ...
#>  $ Vim.7.1  : num  0.004273 0.004273 0.001221 -0.000305 -0.001831 ...
#>  $ Vim.7.2  : num  0.000916 0.003967 0.000916 -0.00061 -0.002136 ...
#>  $ MLC2v.1.1: num  -0.005493 -0.000916 -0.002441 -0.002441 0.00061 ...
#>  $ MLC2v.1.2: num  -0.0061 -0.00305 -0.00305 -0.00153 0.00153 ...
#>  $ MLC2v.2.1: num  -0.005188 -0.00061 -0.002136 -0.00061 0.000916 ...
#>  $ MLC2v.2.2: num  -0.007019 -0.000916 -0.000916 -0.000916 0.00061 ...
#>  $ MLC2v.3.1: num  -0.005493 -0.000916 -0.002441 -0.000916 0.002136 ...
#>  $ MLC2v.3.2: num  -0.00458 0 0 -0.00153 0 ...
#>  $ MLC2v.4.1: num  -0.007019 -0.002441 -0.002441 -0.000916 0.002136 ...
#>  $ MLC2v.4.2: num  -0.0061 -0.00153 -0.00153 -0.00153 0.00153 ...
#>  $ MLC2v.5.1: num  -0.006409 -0.001831 -0.001831 -0.000305 0.001221 ...
#>  $ MLC2v.5.2: num  -0.0058 -0.00122 -0.00122 -0.00122 0.00183 ...
#>  $ MLC2v.6.1: num  -0.006409 -0.001831 -0.001831 -0.000305 0.001221 ...
#>  $ MLC2v.6.2: num  -0.002441 -0.000916 -0.000916 -0.000916 0.002136 ...
#>  $ MLC2v.7.1: num  -0.00458 -0.00153 -0.00305 0 0.00153 ...
#>  $ MLC2v.7.2: num  -0.00153 0 -0.00153 0 0 ...
#>  $ MLC2v.0.1: num  -0.000305 -0.000305 -0.001831 -0.000305 0.001221 ...
#>  $ MLC2v.0.2: num  -0.002136 0.000916 -0.00061 -0.00061 -0.00061 ...