Melt Curves MLC-2v and Vimentin for the qPCR Experiment C60.amp Using the Roche Light Cycler 1.5

Melting curves were continuously monitored with the Light Cycler 1.5 (Roche). For details on sample preparation refer to C60.amp.

data(C60.melt)

Format

A data frame with 128 observations on the following 65 variables. Refer to C60.amp for details of the specific samples.

Index

Index of elements

Vim.0.1.T

Temperature

Vim.0.1.F

Fluorescence

Vim.0.2.T

Temperature

Vim.0.2.F

Fluorescence

Vim.1.1.T

Temperature

Vim.1.1.F

Fluorescence

Vim.1.2.T

Temperature

Vim.1.2.F

Fluorescence

Vim.2.1.T

Temperature

Vim.2.1.F

Fluorescence

Vim.2.2.T

Temperature

Vim.2.2.F

Fluorescence

Vim.3.1.T

Temperature

Vim.3.1.F

Fluorescence

Vim.3.2.T

Temperature

Vim.3.2.F

Fluorescence

Vim.4.1.T

Temperature

Vim.4.1.F

Fluorescence

Vim.4.2.T

Temperature

Vim.4.2.F

Fluorescence

Vim.5.1.T

Temperature

Vim.5.1.F

Fluorescence

Vim.5.2.T

Temperature

Vim.5.2.F

Fluorescence

Vim.6.1.T

Temperature

Vim.6.1.F

Fluorescence

Vim.6.2.T

Temperature

Vim.6.2.F

Fluorescence

Vim.7.1.T

Temperature

Vim.7.1.F

Fluorescence

Vim.7.2.T

Temperature

Vim.7.2.F

Fluorescence

MLC2v.1.1.T

Temperature

MLC2v.1.1.F

Fluorescence

MLC2v.1.2.T

Temperature

MLC2v.1.2.F

Fluorescence

MLC2v.2.1.T

Temperature

MLC2v.2.1.F

Fluorescence

MLC2v.2.2.T

Temperature

MLC2v.2.2.F

Fluorescence

MLC2v.3.1.T

Temperature

MLC2v.3.1.F

Fluorescence

MLC2v.3.2.T

Temperature

MLC2v.3.2.F

Fluorescence

MLC2v.4.1.T

Temperature

MLC2v.4.1.F

Fluorescence

MLC2v.4.2.T

Temperature

MLC2v.4.2.F

Fluorescence

MLC2v.5.1.T

Temperature

MLC2v.5.1.F

Fluorescence

MLC2v.5.2.T

Temperature

MLC2v.5.2.F

Fluorescence

MLC2v.6.1.T

Temperature

MLC2v.6.1.F

Fluorescence

MLC2v.6.2.T

Temperature

MLC2v.6.2.F

Fluorescence

MLC2v.7.1.T

Temperature

MLC2v.7.1.F

Fluorescence

MLC2v.7.2.T

Temperature

MLC2v.7.2.F

Fluorescence

MLC2v.0.1.T

Temperature

MLC2v.0.1.F

Fluorescence

MLC2v.0.2.T

Temperature

MLC2v.0.2.F

Fluorescence

Details

Melting curves were continuously monitored with the Light Cycler 1.5 (Roche).

Source

Stefan Roediger, Claudia Deutschmann (BTU Cottbus - Senftenberg)

References

A Highly Versatile Microscope Imaging Technology Platform for the Multiplex Real-Time Detection of Biomolecules and Autoimmune Antibodies. S. Roediger, P. Schierack, A. Boehm, J. Nitschke, I. Berger, U. Froemmel, C. Schmidt, M. Ruhland, I. Schimke, D. Roggenbuck, W. Lehmann and C. Schroeder. Advances in Biochemical Bioengineering/Biotechnology. 133:33--74, 2013.

Mao, F., Leung, W.-Y., Xin, X., 2007. Characterization of EvaGreen and the implication of its physicochemical properties for qPCR applications. BMC Biotechnol. 7, 76.

Examples

## data(C60.melt)
str(C60.melt)

#> 'data.frame':	128 obs. of  65 variables:
#>  $ Index      : int  0 1 2 3 4 5 6 7 8 9 ...
#>  $ Vim.0.1.T  : num  54.9 55.1 55.3 55.6 55.9 ...
#>  $ Vim.0.1.F  : num  3.41 3.46 3.47 3.47 3.45 ...
#>  $ Vim.0.2.T  : num  54.8 55.1 55.3 55.6 55.9 ...
#>  $ Vim.0.2.F  : num  5.11 5.16 5.18 5.2 5.2 ...
#>  $ Vim.1.1.T  : num  54.8 55.1 55.3 55.6 55.9 ...
#>  $ Vim.1.1.F  : num  5.18 5.18 5.18 5.17 5.16 ...
#>  $ Vim.1.2.T  : num  54.8 55.1 55.3 55.6 55.9 ...
#>  $ Vim.1.2.F  : num  5.29 5.29 5.29 5.28 5.26 ...
#>  $ Vim.2.1.T  : num  54.8 55.1 55.4 55.6 55.9 ...
#>  $ Vim.2.1.F  : num  5.34 5.34 5.34 5.34 5.33 ...
#>  $ Vim.2.2.T  : num  54.8 55.1 55.3 55.6 56 ...
#>  $ Vim.2.2.F  : num  5.51 5.51 5.51 5.5 5.49 ...
#>  $ Vim.3.1.T  : num  54.8 55.1 55.3 55.7 56 ...
#>  $ Vim.3.1.F  : num  5.6 5.6 5.6 5.6 5.58 ...
#>  $ Vim.3.2.T  : num  54.8 55.1 55.3 55.7 56 ...
#>  $ Vim.3.2.F  : num  5.52 5.53 5.53 5.52 5.51 ...
#>  $ Vim.4.1.T  : num  54.8 55.1 55.4 55.7 56 ...
#>  $ Vim.4.1.F  : num  6.85 6.86 6.86 6.86 6.85 ...
#>  $ Vim.4.2.T  : num  54.8 55.1 55.4 55.7 55.9 ...
#>  $ Vim.4.2.F  : num  7.3 7.31 7.31 7.31 7.29 ...
#>  $ Vim.5.1.T  : num  54.8 55.1 55.4 55.7 55.9 ...
#>  $ Vim.5.1.F  : num  5.22 5.23 5.24 5.24 5.24 ...
#>  $ Vim.5.2.T  : num  54.9 55.2 55.4 55.7 55.9 ...
#>  $ Vim.5.2.F  : num  5.39 5.4 5.41 5.4 5.4 ...
#>  $ Vim.6.1.T  : num  54.9 55.2 55.4 55.7 55.9 ...
#>  $ Vim.6.1.F  : num  5.34 5.34 5.35 5.35 5.34 ...
#>  $ Vim.6.2.T  : num  54.9 55.2 55.4 55.7 55.9 ...
#>  $ Vim.6.2.F  : num  5.37 5.37 5.38 5.37 5.37 ...
#>  $ Vim.7.1.T  : num  54.9 55.2 55.4 55.7 55.9 ...
#>  $ Vim.7.1.F  : num  5.15 5.15 5.16 5.15 5.14 ...
#>  $ Vim.7.2.T  : num  54.9 55.2 55.5 55.7 55.9 ...
#>  $ Vim.7.2.F  : num  5.19 5.19 5.2 5.19 5.18 ...
#>  $ MLC2v.1.1.T: num  55 55.2 55.5 55.7 56 ...
#>  $ MLC2v.1.1.F: num  3.96 3.96 3.95 3.94 3.93 ...
#>  $ MLC2v.1.2.T: num  55 55.2 55.5 55.8 55.9 ...
#>  $ MLC2v.1.2.F: num  4.25 4.25 4.24 4.23 4.22 ...
#>  $ MLC2v.2.1.T: num  55 55.2 55.5 55.8 56 ...
#>  $ MLC2v.2.1.F: num  4.13 4.13 4.12 4.12 4.11 ...
#>  $ MLC2v.2.2.T: num  55 55.2 55.5 55.8 56 ...
#>  $ MLC2v.2.2.F: num  4.44 4.44 4.43 4.43 4.41 ...
#>  $ MLC2v.3.1.T: num  55 55.2 55.5 55.8 56 ...
#>  $ MLC2v.3.1.F: num  4.43 4.43 4.42 4.42 4.41 ...
#>  $ MLC2v.3.2.T: num  55 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.3.2.F: num  4.09 4.09 4.08 4.07 4.06 ...
#>  $ MLC2v.4.1.T: num  55 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.4.1.F: num  4.27 4.26 4.26 4.26 4.25 ...
#>  $ MLC2v.4.2.T: num  54.9 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.4.2.F: num  4.01 4.01 4.01 4 3.98 ...
#>  $ MLC2v.5.1.T: num  54.9 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.5.1.F: num  4.01 4.01 4 3.99 3.98 ...
#>  $ MLC2v.5.2.T: num  55 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.5.2.F: num  3.95 3.95 3.94 3.93 3.92 ...
#>  $ MLC2v.6.1.T: num  55 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.6.1.F: num  4.01 4.01 4.01 4 3.99 ...
#>  $ MLC2v.6.2.T: num  55 55.2 55.5 55.8 56.1 ...
#>  $ MLC2v.6.2.F: num  3.82 3.82 3.82 3.8 3.79 ...
#>  $ MLC2v.7.1.T: num  55 55.3 55.5 55.8 56.1 ...
#>  $ MLC2v.7.1.F: num  3.86 3.86 3.87 3.86 3.85 ...
#>  $ MLC2v.7.2.T: num  55 55.3 55.5 55.8 56.1 ...
#>  $ MLC2v.7.2.F: num  3.49 3.49 3.49 3.47 3.46 ...
#>  $ MLC2v.0.1.T: num  55 55.3 55.6 55.9 56.1 ...
#>  $ MLC2v.0.1.F: num  3.12 3.12 3.11 3.08 3.06 ...
#>  $ MLC2v.0.2.T: num  55 55.3 55.6 55.9 56.2 ...
#>  $ MLC2v.0.2.F: num  1.68 1.68 1.68 1.67 1.66 ...