humanrater, a graphical interface to rate curves

humanrater allows rating a curve for a certain characteristic in the interactive, semi-blind manner. humanrater draws individual graphs of a curve and prompts an input field for the user. This function can be used to compare the human rating and the rating of a machine.

humanrater(x, cyc = 1, repeats = 1,
  designations = list(y = "yes", a = "ambiguous", n = "not"),
  shuffle = TRUE, ...)

Arguments

x
is the input data (matrix or data.frame).

cyc
is the index of column containing the cycle data.

repeats
number of repeats to rate the samples.

designations
a named list of length at least 2. See Details.

shuffle
logical, if TRUE sequence of curves is shuffled for purpose of rating.

...
additional arguments to plot function.

Details

A user can specify the list of designations characterizing the curve, where the names of elements specify short designations used during rating. Defaults are y for "yes", a for "ambiguous" and n for "no". The supplied designation list must have length at least two (for example "true"/"false").

Author

Michal Burdukiewicz, Stefan Roediger

Examples

testdata <- data.frame(1:35,
      AmpSim(Cq = 15, noise = TRUE)[, 2],
      AmpSim(Cq = 25, noise = TRUE)[, 2],
      rnorm(35),
      AmpSim(Cq = 35, noise = TRUE)[, 2],
      rnorm(35),
      AmpSim(Cq = 45, noise = TRUE)[, 2])
#we strongly advise against running code below using 'example(humanrater)'
#due to the highly interactive nature of this function (it would not end
#without user's input), it is recommended to just copy lines below into R
#command line
if (FALSE) {
  #check testdata for significance of amplification in two repeats
  human.test1 <- humanrater(testdata, repeats = 2)
  #check testdata for significance of amplification in one repeat and declare more
  #finger friendly (but less obvious) designations
  human.test2 <- humanrater(testdata, repeats = 1, list(q = "yes", w = "no"))
  }