Normalize data

normalizer normalizes any data set using a chosen method (see Details). It may be used when the data from an experiment have considerable variation regarding the background and plateau signal.

normalizer(y, method.norm = "none", qnL = 0.03)

Arguments

y
is a vector containing the fluorescence values.

method.norm
is a argument to use a "none", "minm", "max", "luqn", or "zscore" normalization. See Details.

qnL
is the quantile to be used for the quantile normalization. Ignored if method.norm is not equal to "luqn".

Details

The parameter qnL is a user defined quantile which is used for the quantile normalization. A quantile normalization herein refers to an approach which is less prone to outliers than a normalization based on the minimum and the maximum of an amplification curve. minm does a min-max normalization between 0 and 1 (see Roediger et al. 2013 for explanation). max does a normalization to the maximum value (MFI/max(MFI)). luqn does a quantile normalization based on a symmetric proportion as defined by the qnL parameter (e.g., qnL = 0.03 equals 3 and 97 percent quantiles). zscore performs a z-score normalization with a mean of 0 and a standard deviation of 1.

Value

A vector of normalized fluorescence values.

References

Surface Melting Curve Analysis with R. S. Roediger, A. Boehm and I. Schimke. The R Journal. 5(2):37--52, 2013. https://journal.r-project.org

Author

Stefan Roediger, Michal Burdukiewicz

See also

Examples

normalizer(C17[2L:50, 1], "minm")

#>  [1] 0.00000000 0.02440033 0.04507858 0.06575682 0.08643507 0.10752688
#>  [7] 0.12820513 0.14888337 0.16956162 0.19023987 0.21133168 0.23200993
#> [13] 0.25268817 0.27336642 0.29404467 0.31513648 0.33581472 0.35649297
#> [19] 0.37717122 0.39784946 0.41894127 0.43961952 0.46029777 0.48097601
#> [25] 0.50165426 0.52233251 0.54301075 0.56368900 0.58478081 0.60545906
#> [31] 0.62613730 0.64681555 0.66790736 0.68858561 0.70926385 0.72994210
#> [37] 0.75062035 0.77171216 0.79239041 0.81306865 0.83374690 0.85483871
#> [43] 0.87551696 0.89619520 0.91728701 0.93796526 0.95864351 0.97932175
#> [49] 1.00000000